Optimize array performance
Push is not a very performant operation. It's much faster to create the size of the array ahead of time and then modify each index location. This is akin to C++ std library vectors vs. regular arrays (not exactly, but close analogy).

In some cases direct array access (i.e. [] operator) is an order of magnitude faster. If we were to use Uint8Array's they would be ~3x faster than regular Arrays when comparing direct accesses. TypedArrays do not support the push operation, so there is no comparison there.